feat(sdk): migrate integer fields to use unsigned data types for improved range handling
- Updated various integer fields and parameters to use unsigned equivalents (`u16`, `u32`, `u64`) across SDK entities for better handling of larger values.
- Adjusted serialization and deserialization logic, including method parameters and return types, to account for the unsigned conversions.
- Enhanced type safety and consistency in data processing.

+++ services/src/main/java/systems/glam/services/integrations/kamino/ReserveContext.java
@@ -136,7 +136,7 @@ public record ReserveContext(long slot,
     return tokenInfo.scopeConfiguration();
   }
 
-  public short[] priceChainIndexes() {
+  public int[] priceChainIndexes() {
     return scopeConfiguration().priceChain();
   }
